Refactor to lift Pagination into separate file
